% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftoverGR_hg19ToT2T.R
\name{liftoverGR_hg19ToT2T}
\alias{liftoverGR_hg19ToT2T}
\title{Liftover genomic range from hg19 to t2t (Human genome)}
\usage{
liftoverGR_hg19ToT2T(gr.f)
}
\arguments{
\item{gr.f}{A genomic range.}
}
\value{
Liftover of hg19 genomic loci to t2t genomic loci.
}
\description{
Liftover genomic range from hg19 to t2t (Human genome) and will also add t2t assembly information to the output genomic ranges. (filter out non-standard chromosomes)
}
\examples{
data(hg_38_gr)
outGr_h19=liftoverGR_hg38ToHg19(gr.f=hg_38_gr)
outGr_t2t=liftoverGR_hg19ToT2T(gr.f=outGr_h19)
outGr_t2t


}
