% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftoverGR_hg38ToHg19.R
\name{liftoverGR_hg38ToHg19}
\alias{liftoverGR_hg38ToHg19}
\title{Liftover genomic range from hg38 to hg19 (Human genome)}
\usage{
liftoverGR_hg38ToHg19(gr.f)
}
\arguments{
\item{gr.f}{A genomic range.}
}
\value{
Liftover of hg38 genomic loci to hg19 genomic loci.
}
\description{
Liftover genomic range from hg38 to hg19 (Human genome) and will also add hg19 assembly information to the output genomic ranges. (filter out non-standard chromosomes)
}
\examples{
data(hg_38_gr)
outGr_h19=liftoverGR_hg38ToHg19(gr.f=hg_38_gr)
outGr_h19


}
